\name{corr}
\alias{corr}
\docType{data}
\title{ Correspondence file used as an example by GExMap }
\description{
 Correspondence between ensembl ID and microarray ID.
}
\usage{data(corr)}
\details{
  The data set included in the GExMap package as an example.
}
\source{
http://gexmap.voila.net/index.html
}
\keyword{datasets}
