\name{data}
\alias{data}
\docType{data}
\title{ Genomic data set used as an example by GExMap }
\description{
 Text file tab separated with two columns.
 The first columns is for the probes identifiers  and the second for the variation code (1 or -1).
 The title of the first column must be formated as follow: [type of mocroarray], [microarray name].
 The title of the second column must be "expression".
}
\usage{data(data)}
\details{
  This dat set included in the GExMap package as an example.
}
\source{
http://gexmap.voila.net/index.html
}
\keyword{datasets}
