\name{gexload.corr}
\alias{gexload.corr}
\title{ Function from the GExMap package which annotates identifier list }
\description{
  Creates a list of ENSEMBL id from the user gene list which contains microarray id
}
\usage{
gexload.corr(data, list, scale, source, res)
}
\arguments{
  \item{data}{ The ENSEMBL genome reference data. }
  \item{list}{ The user gene list to be tested and which contains all id which are to be annotated. }
  \item{scale}{ Scale to format the genome. }
  \item{source}{ The source folder where would be placed all the Rdata files. }
  \item{res}{ The results folder where would be saved all the resulint files. }
}
\details{
  This function is included in the GExMap package and could not be used independently.
}
\value{
}
\references{ }
\author{ Nicolas Cagnard <nicolas.cagnard@inserm.fr> }
\note{ 
}
\seealso{ GExMap }
\keyword{ misc }

