\name{gexload.data}
\alias{gexload.data}
\title{ Function from the GExMap package which upload data files to be used by GExMap }
\description{
  Upload automatically (or ask to the user) the data.Rdata file and the *.Rdata file specific to the user identifier type.
}
\usage{
gexload.data(source, genome)
}
\arguments{
  \item{source}{ The source folder where would be placed all the Rdata files. }
  \item{genome}{ ENSEMBL genome whith all genomic informations }
}
\details{
  This function is included in the GExMap package and could not be used independently.
}
\value{
}
\references{  }
\author{ Nicolas Cagnard <nicolas.cagnard@inserm.fr> }
\note{ 
}
\seealso{ GExMap }
\keyword{ misc }

