\name{GExMap-package}
\alias{GExMap-package}
\docType{package}
\title{
GExMap: A visual, intuitive, easy to use software giving access to a new type of information buried into your microarray data.
}
\description{
- Perform statistical tests to unveil genomic clusters\\ 
- Produces garphical interpretations of the statistical results in pdf files\\
- Perform a Gene Ontology analysis and produces graphic results in pdf files\\
}
\details{
\tabular{ll}{
Package: \tab GExMap\cr
Type: \tab Package\cr
Version: \tab 1.1.3\cr
Date: \tab 2011-11-18\cr
License: \tab GPL version 2 or newer\cr
Depends: \tab multtest\cr
LazyLoad: \tab yes\cr
}
}
\author{
Nicolas Cagnard

Maintainer: Nicolas Cagnard <nicolas.cagnard@gmail.com>
}
\references{
http://gexmap.voila.net/index.html
Research Article JPB/Vol.2.1 051- 059(2009)/January 2009 doi:10.4172/jpb.1000060
GExMap: An Intuitive Visual Tool to Detect and Analyze Genomic Distribution in Microarray-generated Lists of Differentially Expressed Genes
Nicolas Cagnard, Carlo Lucchesi, Gilles Chiocchia 
}
