\name{go}
\alias{go}
\title{ GO data file }
\description{
 List of GO ID and descriptions.
}
\details{
  This dat set included in the GExMap package as an example.
}
\value{
}
\references{  }
\author{ Nicolas Cagnard <nicolas.cagnard@inserm.fr> }
\note{
}
\seealso{GExMap}
\keyword{ misc }

