\name{corr}
\alias{corr}
\title{ Correlation file to be used for example by GExMap }
\description{
 Correlation between ensembl ID and microarray ID.
}
\details{
  This dat set included in the GExMap package as an example.
}
\value{
}
\references{  }
\author{ Nicolas Cagnard <nicolas.cagnard@inserm.fr> }
\note{
}
\seealso{GExMap}
\keyword{ misc }

