\name{data}
\alias{data}
\title{ Genomic data set to be used for example by GExMap }
\description{
 Text file tab separated with two columns.
 The first columns is for the probes identifiers  and the second for the variation code (1 or -1).
 The title of the first column must be formated as follow: [type of mocroarray], [microarray name].
 The title of the second column must be "expression".
}
\details{
  This dat set included in the GExMap package as an example.
}
\value{
}
\references{  }
\author{ Nicolas Cagnard <nicolas.cagnard@inserm.fr> }
\note{
}
\seealso{GExMap}
\keyword{ misc }

