\name{gexload.data}
\alias{gexload.data}
\title{ Function from the GExMap package which upload data files to be used by GExMap }
\description{
  Upload automatically (or ask to the user) the data.Rdata file and the *.Rdata file specific to the user identifier type.
}
\usage{
gexload.data(source, genome)
}
\arguments{
  \item{source}{ The source folder where would be placed all the Rdata files. }
  \item{genome}{ ENSEMBL genome whith all genomic informations }
}
\details{
  This function is included in the GExMap package and should not be used independently.
}
\references{
http://gexmap.voila.net/index.html
Research Article JPB/Vol.2.1 051- 059(2009)/January 2009 doi:10.4172/jpb.1000060
GExMap: An Intuitive Visual Tool to Detect and Analyze Genomic Distribution in Microarray-generated Lists of Differentially Expressed Genes
Nicolas Cagnard, Carlo Lucchesi, Gilles Chiocchia 
}
\author{
Nicolas Cagnard <nicolas.cagnard@gmail.com>
}
