\name{go}
\alias{go}
\docType{data}
\title{ GO data file }
\description{
 List of GO ID and descriptions.
}
\usage{data(go)}
\details{
  This dat set included in the GExMap package as an example.
}
\source{
http://gexmap.voila.net/index.html
Research Article JPB/Vol.2.1 051- 059(2009)/January 2009 doi:10.4172/jpb.1000060
GExMap: An Intuitive Visual Tool to Detect and Analyze Genomic Distribution in Microarray-generated Lists of Differentially Expressed Genes
Nicolas Cagnard, Carlo Lucchesi, Gilles Chiocchia 
}
\references{
http://gexmap.voila.net/index.html
}
\keyword{datasets}
