\name{go}
\Rdversion{1.1}
\alias{go}
\docType{data}
\title{
GO annotations data file.
}
\description{
List of all GO terms with IDs and descriptions.
}
\usage{data(go)}
\format{
  Matrix saved in a Rdata file.
}
\details{
Matrix with 3 columns: "id", "name" and "type" of each GO term.
}
\source{
http://www.ensembl.org/biomart/martview
}
\references{
http://gexmap.voila.net/index.html
}
\keyword{datasets}

